Restructure of combat
