+++ combat/advancedCombat/index.js
@@ -1,15 +1,25 @@
 const sleep = require("util").promisify(setTimeout);
 const { generateEmbedCombatRound } = require("./embedGenerator");
-const { getWeaponInfo } = require("./helper");
-const { asyncForEach, randomIntBetweenMinMax } = require("../../game/_GLOBAL_HELPERS");
+const { getWeaponInfo, getRandomWeapons } = require("../../game/_GLOBAL_HELPERS/weapons");
+const { asyncForEach } = require("../../game/_GLOBAL_HELPERS");
+const {
+	combatSetup,
+	checkWinner,
+	validateProgress,
+	handleAdvancedCombatAttack,
+	handleAdvancedCombatHeal,
+	allPlayersAlive } = require("./helper");
+
 /*
 Todo:
 - 'decrypt' npc back to npc form after fight
-- add random weapon or npc weapon when npc fights.
+- Add army options
+- show last embed when fight is over ?
 */
 
 
 const createCombatRound = async (message, progress) => {
+	// Adds all keys and values needed to do combat
 	if (!progress.started) {
 		validateProgress(progress);
 		combatSetup(progress);
@@ -19,21 +29,23 @@ const createCombatRound = async (message, progress) => {
 	}
 
 	// eslint-disable-next-line prefer-const
-	let { numOfAllowedWeapons, allowedWeapons, weaponAnswers } = progress.weaponInformation;
+	let { numOfAllowedWeapons, weaponAnswers } = progress.weaponInformation;
 
-	allowedWeapons = getWeaponInfo(null, numOfAllowedWeapons);
+	// Gets x random weapons
+	const allowedWeapons = getRandomWeapons(numOfAllowedWeapons);
 	progress.weaponInformation.allowedWeapons = allowedWeapons;
 
 	const weaponAnswerFilter = Object.keys(allowedWeapons)
 		.map((w) => [allowedWeapons[w].answer, allowedWeapons[w].name])
 		.flat();
 
+
 	const combatRound = generateEmbedCombatRound(progress);
 	await message.channel.send(combatRound);
 
 	const filter = (response) => {
-		// checks if included in the fight
-		return progress.teamGreenIds.includes(response.author.id) || progress.teamRedIds.includes(response.author.id);
+		// checks if person included included in the fight
+		return [...progress.teamGreenIds, ...progress.teamRedIds].includes(response.author.id);
 	};
 
 	const collector = await message.channel.createMessageCollector(filter, {
@@ -52,7 +64,7 @@ const createCombatRound = async (message, progress) => {
 		}
 		const answer = result.content.toLowerCase();
 
-		// adds answer to progress object
+		// Allows user to type in a,b,c OR name of the weapon eg 'slash'
 		if (Object.keys(allowedWeapons).includes(answer)) {
 			weaponAnswers.set(result.author.id, answer);
 		}
@@ -62,8 +74,8 @@ const createCombatRound = async (message, progress) => {
 			);
 			weaponAnswers.set(result.author.id, weaponInformation.name);
 		}
-		// stops collecting if all users have answered
-		if (weaponAnswers.size >= progress.teamRedIds.length + progress.teamGreenIds.length) {
+		// stops collecting if all humans have answered
+		if (weaponAnswers.size >= [...progress.teamGreen, ...progress.teamRed].filter(player=> !player.account.npc).length) {
 			await sleep(1500);
 			collector.stop();
 		}
@@ -75,6 +87,7 @@ const createCombatRound = async (message, progress) => {
 				resolve(progress);
 			}
 			else {
+				// todo show last embed
 				resolve(await createCombatRound(message, result));
 			}
 		});
@@ -89,41 +102,46 @@ const calculateCombatResult = async (progress) => {
 	// cleans up roundResults from previous round
 	progress.roundResults = [];
 
+	// Stores all actions in the variables before saving to database or npc-object
 	const awaitHealPlayerPromises = {};
 	const awaitDamagePlayerPromises = {};
 
 	// adds weapon choice from npc
-	// clean this
 	[...teamRed, ...teamGreen]
 		.filter(player=>player.account.npc)
 		.forEach(player=>{
-			const weaponInformation = Object.values(progress.weaponInformation.allowedWeapons).find(
-				(weapon) => weapon.answer === "a"
-			);
-			weaponAnswers.set(player.account.userId, weaponInformation.name);
+			let weaponName;
+			// chooses a random weapon from npc personal arsenal if it exists
+			if (player.weapons && player.weapons.length) {
+				weaponName = player.weapons[Math.floor(Math.random() * player.weapons.length)];
+			}
+			// OR chooses a random weapon from the current allowed weapons
+			else {
+				weaponName = Object.keys(progress.weaponInformation.allowedWeapons)[Math.floor(Math.random() * Object.keys(progress.weaponInformation.allowedWeapons).length)];
+			}
+			// Saves the answer to the weaponanswer
+			weaponAnswers.set(player.account.userId, weaponName);
 		});
 
 
 	// loops through every weaponanswer and performs action
-	weaponAnswers.forEach((weapon, player) => {
+	weaponAnswers.forEach((weapon, playerId) => {
 		// allows players with more attack to attack more than once
-		const allowedNumOfAttacks = player.allowedNumOfAttacks || 1;
+		const isTeamGreen = teamGreen.some(member=> member.account.userId === playerId);
+
+		// Figure out which team is friendly and which is opposing
+		const friendlyTeam = isTeamGreen ? teamGreen : teamRed;
+		const opposingTeam = isTeamGreen ? teamRed : teamGreen;
+
+		const playerInfo = friendlyTeam.find((member) => member.account.userId === playerId);
+		// allows for multiple attack
+		const allowedNumOfAttacks = playerInfo.allowedNumOfAttacks || 1;
 		for (let i = 0; i < allowedNumOfAttacks; i += 1) {
-
-			const isTeamGreen = teamGreen.some(member=> member.account.userId === player);
-
-			// Figure out which team is friendly and which is opposing
-			const friendlyTeam = isTeamGreen ? teamGreen : teamRed;
-			const opposingTeam = isTeamGreen ? teamRed : teamGreen;
-
-			const playerInfo = friendlyTeam.find((member) => member.account.userId === player);
-			const randomVictimInfo = opposingTeam[Math.floor(Math.random() * teamRed.length)];
+			const randomVictimInfo = opposingTeam[Math.floor(Math.random() * opposingTeam.length)];
 
 			// lower chance is better
 			const chance = Math.random();
 			const weaponInfo = getWeaponInfo(weapon);
-			const playerName = playerInfo.account.username;
-			const victimName = randomVictimInfo.account.username;
 
 			if (weaponInfo.chanceforSuccess > chance) {
 				if (weaponInfo.type === "attack") {
@@ -134,7 +152,9 @@ const calculateCombatResult = async (progress) => {
 				}
 			}
 			else {
-				progress.roundResults.push(`\n${playerName} failed to use ${weaponInfo.name} on ${victimName === playerName ? "himself" : victimName}`);
+				const playerName = playerInfo.account.username;
+				const victimName = randomVictimInfo.account.username;
+				progress.roundResults.push(`\n**${playerName}** failed to use ${weaponInfo.name} on **${victimName === playerName ? "himself" : victimName}**`);
 			}
 		}
 	});
@@ -155,9 +175,7 @@ const calculateCombatResult = async (progress) => {
 	// removes player from combat
 	teamGreen = teamGreen.filter(member=>member.hero.currentHealth > 0);
 	teamRed = teamRed.filter(member=>member.hero.currentHealth > 0);
-	console.log(progress.currentRound);
 	progress.currentRound += 1;
-	console.log(progress.currentRound, "progress.currentRound");
 	progress.weaponInformation.weaponAnswers.clear();
 
 	// checks if fight is over
@@ -165,188 +183,5 @@ const calculateCombatResult = async (progress) => {
 	return progress;
 };
 
-const generateAttackString = (playerName, weaponInfo, damageGiven, playerAttacked) => {
-	return `\n- **${playerName}** used ${weaponInfo.name} attack causing **${damageGiven}** damage to **${playerAttacked}**`;
-
-};
-const generateHealString = (playerName, weaponInfo, healGiven, playerHealed) => {
-	return `\n${playerName} helead ${playerHealed === playerName ? "himself" : playerHealed}. +${healGiven} HP`;
-};
-
-const combatSetup = progress => {
-	setupProgressKeys(progress);
-	convertNpcsToHuman(progress);
-	populateDiscordIds(progress);
-	populatePlayerNames(progress);
-	formatEmbedInformation(progress);
-};
-
-const formatEmbedInformation = (progress)=> {
-	const { embedInformation } = progress;
-	embedInformation.teamRedName = embedInformation.teamRedName || "Team Red";
-	embedInformation.teamGreenName = embedInformation.teamGreenName || "Team Green";
-	embedInformation.title = embedInformation.title || "BATTLE!";
-	embedInformation.description = embedInformation.description || "Here is description";
-	embedInformation.fields = embedInformation.fields || [];
-	embedInformation.footer = embedInformation.footer || "TIP: Write your weapon of choice in the chat. eg -> a or c";
-
-	return progress;
-
-};
-
-const convertNpcsToHuman = (progress) => {
-	progress.teamRed.map(member=> {
-		if (!member.account) return convertNpc(member);
-	});
-	progress.teamGreen.map(member=> {
-		if (!member.account) return convertNpc(member);
-	});
-};
-
-const convertNpc = (npc)=> {
-	npc.account = {
-		userId: Math.random().toString(36).substr(2, 10),
-		username: npc.name,
-		npc: true,
-	};
-	npc.hero = {
-		health: npc.stats.health,
-		currentHealth: npc.stats.health,
-		attack: npc.stats.attack
-	};
-
-	npc.heroHpLossFixedAmount = function(hp) {
-		this.currentHealth -= hp;
-	};
-	npc.healHero = function(hp) {
-		this.currentHealth += hp;
-	};
-
-	npc.stats = null;
-	npc.name = null;
-	return npc;
-
-};
-
-
-const setupProgressKeys = (progress)=>{
-	const setup = {
-		started: true,
-		winner: null,
-		roundResults: [],
-		currentRound: 0,
-		weaponInformation: {
-			numOfAllowedWeapons: 3,
-			allowedWeapons: null,
-			weaponAnswers: new Map,
-		},
-		teamGreenIds:[],
-		teamRedIds: [],
-		teamGreenNames: [],
-		teamRedNames: [],
-	};
-	Object.assign(progress, setup);
-};
-
-
-const populateDiscordIds = (progress)=>{
-	progress.teamGreen.forEach(member=> progress.teamGreenIds.push(member.account.userId));
-	progress.teamRed.forEach(member=> progress.teamRedIds.push(member.account.userId));
-};
-
-const populatePlayerNames = (progress)=>{
-	progress.teamGreen.forEach(member=> progress.teamGreenNames.push(member.account.username));
-	progress.teamRed.forEach(member=> progress.teamRedNames.push(member.account.username));
-};
-
-const checkWinner = progress=>{
-	const{ teamGreen, teamRed } = progress;
-	if (teamGreen.length === 0 && teamRed.length === 0) {
-		return "draw";
-	}
-	if (teamGreen.length === 0 || teamRed.length === 0) {
-		return teamGreen.length === 0 ? "teamRed" : "teamGreen";
-	}
-	if (progress.currentRound > progress.combatRules.maxRounds) {
-		return "No winners";
-	}
-	return null;
-};
-
-const validateProgress = (progress)=>{
-
-	const progressKeys = ["combatRules", "teamGreen", "teamRed", "embedInformation"];
-	if (!progressKeys.some(key=> Object.keys(progress).includes(key))) {
-		throw new Error(`progress keys are missing\nExpected: ${progressKeys}\nGot: ${Object.keys(progress)}\n`);
-	}
-	if (!progress.combatRules.maxRounds || typeof progress.combatRules.maxRounds !== "number") {
-		throw new Error("progress.combatRules.maxRounds must be set to a number\n");
-	}
-	if (progress.combatRules.armyAllowed === undefined) {
-		throw new Error("progress.combatRules.armyAllowed must be set to a boolean\n");
-	}
-	if (progress.teamGreen.length === 0 || progress.teamRed.length === 0) {
-		throw new Error(`No players in the teams. \n teamGreen: ${progress.teamGreen.length} members \n teamRed: ${progress.teamRed.length} members\n`);
-	}
-};
-
-const handleAdvancedCombatAttack = (playerInfo, weaponInfo, awaitDamagePlayerPromises, randomVictimInfo, progress) =>{
-	const victimName = randomVictimInfo.account.username;
-	const playerName = playerInfo.account.username;
-	const damageGiven = randomIntBetweenMinMax(
-		(playerInfo.hero.attack / 2) * weaponInfo.damage,
-		playerInfo.hero.attack * weaponInfo.damage
-	);
-	if (awaitDamagePlayerPromises[victimName]) {
-		awaitDamagePlayerPromises[victimName].damage += damageGiven;
-	}
-	else {
-		awaitDamagePlayerPromises[victimName] = {
-			user: randomVictimInfo,
-			damage: damageGiven,
-		};
-	}
-	progress.roundResults.push(generateAttackString(playerName, weaponInfo, damageGiven, victimName)
-	);
-
-};
-
-const handleAdvancedCombatHeal = (playerInfo, friendlyTeam, weaponInfo, awaitHealPlayerPromises, progress)=>{
-	const playerName = playerInfo.account.username;
-	const teamMateWithLowestHp = friendlyTeam
-		.sort((a, b)=> a.hero.currentHealth - b.hero.currentHealth)
-		.filter((p)=> p.hero.currentHealth > 0)[0];
-	const teamMateName = teamMateWithLowestHp.account.username;
-
-	const playerHealedName = teamMateWithLowestHp.account.username;
-	const healGiven = randomIntBetweenMinMax(
-		(playerInfo.hero.health * weaponInfo.damage) / 2,
-		playerInfo.hero.health * weaponInfo.damage
-	);
-	if (awaitHealPlayerPromises[teamMateName]) {
-		awaitHealPlayerPromises[teamMateName].healGiven += healGiven;
-	}
-	else {
-		awaitHealPlayerPromises[teamMateName] = {
-			user: teamMateWithLowestHp,
-			healGiven: healGiven,
-		};
-	}
-	progress.roundResults.push(
-		generateHealString(
-			playerName,
-			weaponInfo,
-			healGiven,
-			playerHealedName
-		)
-	);
-};
-
-const allPlayersAlive = progress => {
-	const { teamGreen, teamRed } = progress;
-	return [...teamGreen, ...teamRed].every(player=>{
-		return player.hero.currentHealth > 0;
-	});
-};
 
 module.exports = { createCombatRound };
