+++ game/tower/army-tower/army-tower.js
@@ -33,7 +33,6 @@ const armyTowerFight = async (users, category, message) => {
 		enemy.stats[stat] = Math.floor(enemy.stats[stat] * enemyCombatModifier);
 	}
 
-	// console.log(users, "user");
 	const templateProgress = {
 		combatRules:{
 			armyAllowed: false,
@@ -45,7 +44,6 @@ const armyTowerFight = async (users, category, message) => {
 	};
 
 	const combatResults = await createCombatRound(message, templateProgress);
-	console.log("done");
 
 	// const combatResults = users.map(user => calculatePveFullArmyResult(user, enemy));
 
@@ -161,9 +159,7 @@ const armyTowerFight = async (users, category, message) => {
 // Takes the user and returns true if it still has health left on some units
 const healthLeftOnArmy = (user) => {
 	const totalUnitsArray = Object.values(user.army.units).map(unitBuild => Object.values(unitBuild)).flat();
-
 	const totalUnits = totalUnitsArray.filter(unitNumbers => typeof unitNumbers === "number").reduce((acc, cur) => acc + cur);
-
 	return totalUnits > 0 || user.hero.currentHealth > 0;
 };
